Create 'std/mat.ama' module and setup import paths
