+++ amanda/vm/src/modules/builtins/mod.rs
@@ -1,6 +1,7 @@
 use crate::ama_value::AmaValue;
 use crate::errors::AmaErr;
 mod embutidos;
+mod mat;
 mod utils;
 
 /*Helpers*/
@@ -9,5 +10,6 @@ pub(crate) type BuiltinDefs<'a> =
     rustc_hash::FxHashMap<&'a str, rustc_hash::FxHashMap<&'a str, AmaValue<'a>>>;
 
 utils::builtin_defs! {
-    embutidos
+    embutidos,
+    mat
 }
