+++ tests/test_cases/math/math_builtins.ama
@@ -1,3 +1,5 @@
+usa "mat";
+
 escrevaln(arredonda(expoente(9, 2)))
 escrevaln(abs(-9))
 escrevaln(abs(9))
